+++ src/snakemake/dag.py
@@ -15,7 +15,7 @@ import tarfile
 import textwrap
 import time
 import json
-from typing import Iterable, List, Optional, Set, Union
+from typing import Iterable, List, Mapping, Optional, Set, Union, Dict
 import uuid
 from collections import Counter, defaultdict, deque, namedtuple
 from functools import partial
@@ -31,6 +31,7 @@ from snakemake.settings.types import DeploymentMethod
 from snakemake_interface_executor_plugins.dag import DAGExecutorInterface
 from snakemake_interface_report_plugins.interfaces import DAGReportInterface
 from snakemake_interface_storage_plugins.storage_object import StorageObjectTouch
+from snakemake_interface_scheduler_plugins.interfaces.dag import DAGSchedulerInterface
 from snakemake_interface_logger_plugins.common import LogEvent
 from snakemake.settings.enums import Quietness
 
@@ -101,7 +102,7 @@ def toposort(graph):
     return sorted
 
 
-class DAG(DAGExecutorInterface, DAGReportInterface):
+class DAG(DAGExecutorInterface, DAGReportInterface, DAGSchedulerInterface):
     """Directed acyclic graph of jobs."""
 
     def __init__(
@@ -122,8 +123,11 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
         ignore_incomplete=False,
         rules_allowed_for_needrun: AnySet[str] = frozenset(),
     ):
+        self._deferred_temp_jobs = []
         self._queue_input_jobs = None
-        self._dependencies = defaultdict(partial(defaultdict, set))
+        self._dependencies: Mapping[Job, Mapping[Job, Set[str]]] = defaultdict(
+            partial(defaultdict, set)
+        )
         self.depending = defaultdict(partial(defaultdict, set))
         self._needrun = set()
         self._checkpoint_jobs = set()
@@ -143,7 +147,7 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
         self.priorityfiles = priorityfiles
         self.priorityrules = priorityrules
         self.targetjobs = set()
-        self.derived_targetfiles = None
+        self._derived_targetfiles = None
         self.prioritytargetjobs = set()
         self._ready_jobs = set()
         self._jobid = dict()
@@ -158,6 +162,8 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
         self._storage_input_jobs = defaultdict(list)
         self.max_checksum_file_size = self.workflow.dag_settings.max_checksum_file_size
         self._checked_jobs = set()
+        self._checked_needrun_jobs = set()
+        self._seen_outputs: Dict[str, Union[Job, GroupJob]] = dict()
 
         self.job_factory = JobFactory()
         self.group_job_factory = GroupJobFactory()
@@ -198,8 +204,20 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
         self.update_output_index()
 
     @property
-    def dependencies(self):
+    def derived_targetfiles(self):
+        if self._derived_targetfiles is None:
+            self._derived_targetfiles = {
+                f for job in self.targetjobs if not job.output for f in job.input
+            } | self.targetfiles
+        return self._derived_targetfiles
+
+    @property
+    def dependencies(self) -> Mapping[Job, Mapping[Job, Set[str]]]:
         return self._dependencies
+
+    def job_dependencies(self, job: Job) -> Iterable[Job]:
+        # keys() returns an iterable, but somehow the type checker does not recognize it
+        return self._dependencies.get(job, {}).keys()  # type: ignore[return-value]
 
     @property
     def batch(self):
@@ -245,10 +263,6 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
             self.targetjobs.add(job)
             self.forcefiles.update(job.output)
 
-        self.derived_targetfiles = {
-            f for job in self.targetjobs if not job.output for f in job.input
-        } | self.targetfiles
-
         self.cleanup()
 
         await self.check_incomplete()
@@ -272,10 +286,6 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
 
         self.check_directory_outputs()
 
-        # check if remaining jobs are valid
-        for i, job in enumerate(self.jobs):
-            job.is_valid()
-
     def get_unneeded_temp_files(self, job: AbstractJob) -> Iterable[str]:
         if isinstance(job, GroupJob):
             for j in job:
@@ -288,7 +298,7 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
     def check_directory_outputs(self):
         """Check that no output file is contained in a directory output of the same or another rule."""
         outputs = sorted(
-            {(os.path.abspath(f), job) for job in self.jobs for f in job.output}
+            {(Path(f).absolute(), job) for job in self.jobs for f in job.output}
         )
         for i in range(len(outputs) - 1):
             (a, job_a), (b, job_b) = outputs[i : i + 2]
@@ -601,7 +611,7 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
         """All jobs in the DAG."""
         return self._dependencies.keys()
 
-    def needrun_jobs(self, exclude_finished=True):
+    def needrun_jobs(self, exclude_finished=True) -> Iterable[Job]:
         """Jobs that need to be executed."""
         if exclude_finished:
             return filterfalse(self.finished, self._needrun)
@@ -642,7 +652,7 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
         """Return the reason of the job execution."""
         return self._reason[job]
 
-    def finished(self, job):
+    def finished(self, job) -> bool:
         """Return whether a job is finished."""
         return job in self._finished
 
@@ -917,12 +927,6 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
                 for f in files:
                     if f in job_.temp_output and f not in skip:
                         yield f
-
-    async def temp_size(self, job):
-        """Return the total size of temporary input files of the job.
-        If none, return 0.
-        """
-        return sum([await f.size() for f in self.temp_input(job)])
 
     def is_needed_tempfile(self, job, tempfile):
         return (
@@ -1448,15 +1452,15 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
                             # for determining any other changes than file modification dates, as it will
                             # change after evaluating the input function of the job in the second pass.
 
-                            # The list comprehension is needed below in order to
-                            # collect all the async generator items before
-                            # applying any().
-                            reason.code_changed = any(
-                                [
-                                    f
-                                    async for f in job.outputs_older_than_script_or_notebook()
-                                ]
-                            )
+                            if RerunTrigger.CODE in self.workflow.rerun_triggers:
+                                # Always check script/notebook mtime unless CODE trigger is disabled.
+                                # Short-circuit on first older output to avoid building a list.
+                                reason.code_changed = False
+                                async for (
+                                    _
+                                ) in job.outputs_older_than_script_or_notebook():
+                                    reason.code_changed = True
+                                    break
                             if not self.workflow.persistence.has_metadata(job):
                                 reason.no_metadata = True
                             elif self.workflow.persistence.has_outdated_metadata(job):
@@ -1897,7 +1901,7 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
 
         self.update_ready()
 
-        await self.check_needrun_jobs()
+        await self.check_jobs()
 
         if check_initial:
             assert not (not self.ready_jobs and any(self.needrun_jobs())), (
@@ -1905,9 +1909,32 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
                 "ready for execution."
             )
 
-    async def check_needrun_jobs(self):
-        for job in filterfalse(self._checked_jobs.__contains__, self.needrun_jobs()):
+    async def check_jobs(self):
+        # first we check all **needrun** jobs whether its output can be made
+        for job in filterfalse(
+            self._checked_needrun_jobs.__contains__, self.needrun_jobs()
+        ):
             await job.check_protected_output()
+            self._checked_needrun_jobs.add(job)
+
+        # now we check **all* jobs for validity
+        for job in filterfalse(self._checked_jobs.__contains__, self.jobs):
+            job.is_valid()
+
+            # here we check if no two rules make the same output
+            # this can happen in edge cases where the output of jobs
+            # are not part of the targetfiles
+            for output_file in job.output:
+                if output_file in self._seen_outputs:
+                    other_job = self._seen_outputs[output_file]
+                    if other_job.jobid == job.jobid:
+                        continue
+
+                    if not self.workflow.execution_settings.ignore_ambiguity:
+                        raise AmbiguousRuleException(output_file, other_job, job)
+                else:
+                    self._seen_outputs[output_file] = job
+
             self._checked_jobs.add(job)
 
     def handle_pipes_and_services(self):
@@ -2174,8 +2201,9 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
             or self.workflow.remote_exec
         ):
             await self.retrieve_storage_inputs()
+        self._derived_targetfiles = None
 
-    def register_running(self, jobs):
+    def register_running(self, jobs: AnySet[AbstractJob]):
         self._running.update(jobs)
         self._ready_jobs -= jobs
         for job in jobs:
@@ -2246,13 +2274,16 @@ class DAG(DAGExecutorInterface, DAGReportInterface):
                 self.create_conda_envs()
             potential_new_ready_jobs = True
 
-        if not self.checkpoint_jobs:
+        if self.checkpoint_jobs:
             # While there are still checkpoint jobs, we cannot safely delete
             # temp files.
             # TODO: we maybe could be more accurate and determine whether there is a
             # checkpoint that depends on the temp file.
-            for job in jobs:
+            self._deferred_temp_jobs.extend(jobs)
+        else:
+            for job in chain(jobs, self._deferred_temp_jobs):
                 await self.handle_temp(job)
+            self._deferred_temp_jobs.clear()
 
         return potential_new_ready_jobs